+++ src/components/common/profile/DefaultMyPageProfileImage.tsx
@@ -1,19 +1,17 @@
 import React, { ChangeEvent, useRef } from 'react';
 import { useMutation } from '@tanstack/react-query';
 import uploadProfileImage from '@/api/uploadProfileImage';
+import { DefaultMyPageProfileImageProps } from '@/types/myPageProfile';
+import { useLocation } from 'react-router-dom';
 
 const DefaultMyPageProfileImage = ({
   nickname,
   setUploadedImage = () => null,
   setIsShowDefaultImage,
-}: {
-  nickname: string;
-  setUploadedImage: React.Dispatch<React.SetStateAction<string | null>>;
-  setIsShowDefaultImage: React.Dispatch<React.SetStateAction<boolean>>;
-}) => {
+}: DefaultMyPageProfileImageProps) => {
   const nicknameInitial = nickname[0];
   const fileInputRef = useRef<HTMLInputElement>(null);
-
+  const location = useLocation();
   const mutation = useMutation({
     mutationFn: uploadProfileImage,
     onSuccess: (data) => {
@@ -44,12 +42,15 @@ const DefaultMyPageProfileImage = ({
     <div className="relative w-40 h-40 bg-slate-400 rounded-full flex items-center justify-center text-white">
       <span className="text-[80px]">{nicknameInitial}</span>
 
-      <div
-        onClick={handleFileInputClick}
-        className="absolute p-[10px] w-11 h-11 inline-flex items-start top-[115px] right-3 z-10 rounded-full bg-green-80 cursor-pointer"
-      >
-        <img className="w-6 h-6" src="/assets/pen_icon.svg" alt="penIcon" />
-      </div>
+      {location.pathname === '/my/profile' && (
+        <button
+          type="button"
+          onClick={handleFileInputClick}
+          className="absolute p-[10px] w-11 h-11 inline-flex items-start top-[115px] right-3 z-10 rounded-full bg-green-80 cursor-pointer"
+        >
+          <img className="w-6 h-6" src="/assets/pen_icon.svg" alt="penIcon" />
+        </button>
+      )}
 
       <input
         ref={fileInputRef}
