+++ src/components/common/profile/MyPageProfile.tsx
@@ -1,8 +1,12 @@
 import { useLocation } from 'react-router-dom';
 import { MyPageProfileProps } from '@/types/myPageProfile';
+import uploadProfileImage from '@/api/uploadProfileImage';
+
 import useUserInfoQuery from '@/hooks/useUserInfoQuery';
-import PageMenu from './PageMenu';
-import ProfileImage from './ProfileImage';
+import React, { ChangeEvent } from 'react';
+import DefaultMyPageProfileImage from './DefaultMyPageProfileImage';
+import MyPageProfileWithUrl from './MyPageProfileWithUrl';
+import ProfileImageLayout from './ProfileImageLayout';
 
 const MyPageProfile = ({
   uploadedImage = null,
@@ -13,6 +17,7 @@ const MyPageProfile = ({
   setIsShowDefaultImage,
 }: MyPageProfileProps) => {
   const location = useLocation();
+
   const isPathNameMyProfile = location.pathname === '/my/profile';
   const isShowProfile = isPathNameMyProfile && !isShowProfileForm;
 
@@ -26,51 +31,59 @@ const MyPageProfile = ({
     return <div>프로필을 불러오는데 실패했습니다</div>;
   }
 
-  return (
-    <div
-      className={`flex ${isShowProfile ? '' : 'sm:hidden'} md:w-[251px] lg:w-96 p-6 flex-col justify-center items-start gap-6 border rounded-xl border-gray-50 bg-white shadow-md self-start`}
-    >
-      <div className="flex justify-center items-start gap-[227px] self-stretch">
-        <div className="flex flex-col justify-center items-center gap-6">
-          <ProfileImage
-            nickname={userInfo.nickname}
-            profileImageUrl={userInfo.profileImageUrl}
-            uploadedImage={uploadedImage}
-            setUploadedImage={setUploadedImage}
-            isShowDefaultImage={isShowDefaultImage}
-            setIsShowDefaultImage={setIsShowDefaultImage}
-          />
-        </div>
-      </div>
+  const handleFileChange = async (e: ChangeEvent<HTMLInputElement>) => {
+    const file = e.target.files?.[0];
+    setIsShowDefaultImage(false);
+    if (file) {
+      try {
+        const imageData = await uploadProfileImage(file);
+        setUploadedImage(imageData.profileImageUrl);
+      } catch (error) {
+        console.error('Failed to upload image:', error);
+      }
+    }
+  };
 
-      <div className=" flex flex-col items-start gap-2 self-stretch">
-        <PageMenu
-          linkTo="/my/profile"
-          icon="/assets/account_check_icon_gray.svg"
-          activeIcon="/assets/account_check_icon.svg"
-          name="내 정보"
-          setIsShowProfileForm={setIsShowProfileForm}
+  const handleUploadImage = () => {
+    const fileInput = document.getElementById('file-input');
+    if (fileInput) fileInput.click();
+  };
+
+  const handleDeleteProfileImage = () => {
+    setUploadedImage(null);
+    setIsShowDefaultImage(true);
+  };
+
+  if (!userInfo.profileImageUrl && !uploadedImage) {
+    return (
+      <ProfileImageLayout isShowProfile={isShowProfile} setIsShowProfileForm={setIsShowProfileForm}>
+        <DefaultMyPageProfileImage
+          nickname={userInfo.nickname}
+          setUploadedImage={setUploadedImage}
+          setIsShowDefaultImage={setIsShowDefaultImage}
         />
-        <PageMenu
-          linkTo="/my/reservation-history"
-          icon="/assets/textbox_check_icon_gray.svg"
-          activeIcon="/assets/textbox_check_icon.svg"
-          name="예약 내역"
+      </ProfileImageLayout>
+    );
+  }
+
+  return (
+    <ProfileImageLayout isShowProfile={isShowProfile} setIsShowProfileForm={setIsShowProfileForm}>
+      {isShowDefaultImage === true ? (
+        <DefaultMyPageProfileImage
+          nickname={userInfo.nickname}
+          setUploadedImage={setUploadedImage}
+          setIsShowDefaultImage={setIsShowDefaultImage}
         />
-        <PageMenu
-          linkTo="/my/activity"
-          icon="/assets/setting_icon_gray.svg"
-          activeIcon="/assets/setting_icon.svg"
-          name="내 체험 관리"
+      ) : (
+        <MyPageProfileWithUrl
+          uploadedImage={uploadedImage}
+          profileImageUrl={userInfo.profileImageUrl}
+          handleFileChange={handleFileChange}
+          handleUploadImage={handleUploadImage}
+          handleDeleteProfileImage={handleDeleteProfileImage}
         />
-        <PageMenu
-          linkTo="/my/reservation-status"
-          icon="/assets/calendar_check_icon_gray.svg"
-          activeIcon="/assets/calendar_check_icon.svg"
-          name="예약 현황"
-        />
-      </div>
-    </div>
+      )}
+    </ProfileImageLayout>
   );
 };
 export default MyPageProfile;
