+++ src/types/myPageProfile.ts
@@ -1,6 +1,8 @@
+import { ChangeEvent, ReactNode } from 'react';
+
 export interface MyPageProfileProps {
-  uploadedImage?: string | null;
-  setUploadedImage?: React.Dispatch<React.SetStateAction<string | null>>;
+  uploadedImage: string | null;
+  setUploadedImage: React.Dispatch<React.SetStateAction<string | null>>;
   isShowProfileForm: boolean;
   setIsShowProfileForm: React.Dispatch<React.SetStateAction<boolean>>;
   isShowDefaultImage: boolean;
@@ -10,8 +12,8 @@ export interface MyPageProfileProps {
 export interface ProfileImageProps {
   nickname: string;
   profileImageUrl: string;
-  uploadedImage?: string | null;
-  setUploadedImage?: React.Dispatch<React.SetStateAction<string | null>>;
+  uploadedImage: string | null;
+  setUploadedImage: React.Dispatch<React.SetStateAction<string | null>>;
   isShowDefaultImage: boolean;
   setIsShowDefaultImage: React.Dispatch<React.SetStateAction<boolean>>;
 }
@@ -22,4 +24,24 @@ export interface PageMenuProps {
   activeIcon: any;
   name: string;
   setIsShowProfileForm?: React.Dispatch<React.SetStateAction<boolean>>;
+}
+
+export interface MyPageProfileWithUrlProps {
+  profileImageUrl: string;
+  uploadedImage: string | null;
+  handleFileChange: (e: ChangeEvent<HTMLInputElement>) => Promise<void>;
+  handleUploadImage: () => void;
+  handleDeleteProfileImage: () => void;
+}
+
+export interface DefaultMyPageProfileImageProps {
+  nickname: string;
+  setUploadedImage?: React.Dispatch<React.SetStateAction<string | null>>;
+  setIsShowDefaultImage: React.Dispatch<React.SetStateAction<boolean>>;
+}
+
+export interface ProfileImageLayoutProps {
+  children: ReactNode;
+  isShowProfile: boolean;
+  setIsShowProfileForm: React.Dispatch<React.SetStateAction<boolean>>;
 }
